replace portfolio page with the revamped portfolios page
