+++ src/components/dashboard-portfolio-section/dashboard-portfolio-section.tsx
@@ -11,14 +11,39 @@ const DashboardPortfolioSection: FC<IDashboardPortfolioSectionProps> = ({
   onPortfolioSelect
 }) => (
   <section className="outline-test flex h-full w-full items-center justify-center">
-    <div className="flex w-full grid-cols-4 flex-col gap-3 px-6 lg:grid">
+    <div className="flex w-full grid-cols-4 flex-col gap-3 lg:grid">
       <div className="col-span-3 max-w-full">
-        <section className="grid h-full auto-cols-[333px] grid-flow-col gap-3">
+        <section className="grid h-full auto-cols-[250px] grid-flow-col gap-3">
           {portfolios.map(portfolio => (
             <PortfolioOverviewCard
               key={portfolio._id}
               portfolio={portfolio}
               onPortfolioSelect={onPortfolioSelect}
+              variant="compact"
+            />
+          ))}
+          {portfolios.map(portfolio => (
+            <PortfolioOverviewCard
+              key={portfolio._id}
+              portfolio={portfolio}
+              onPortfolioSelect={onPortfolioSelect}
+              variant="compact"
+            />
+          ))}
+          {portfolios.map(portfolio => (
+            <PortfolioOverviewCard
+              key={portfolio._id}
+              portfolio={portfolio}
+              onPortfolioSelect={onPortfolioSelect}
+              variant="compact"
+            />
+          ))}
+          {portfolios.map(portfolio => (
+            <PortfolioOverviewCard
+              key={portfolio._id}
+              portfolio={portfolio}
+              onPortfolioSelect={onPortfolioSelect}
+              variant="compact"
             />
           ))}
         </section>
