+++ src/components/portfolio-widgets/portfolio-section/portfolio-section.tsx
@@ -13,7 +13,7 @@ const PortfolioSection = ({
 }) => (
   <article className="mb-6">
     <h1 className="mb-4 text-3xl font-semibold">{title}</h1>
-    <div className="grid grid-cols-1 gap-4 p-2 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4">
+    <div className="grid auto-cols-[250px] grid-flow-col gap-5 p-2">
       {portfolios
         .filter(portfolio => portfolio.assetType === assetType)
         .map(portfolio => (
@@ -21,6 +21,7 @@ const PortfolioSection = ({
             key={portfolio._id}
             portfolio={portfolio}
             onPortfolioSelect={setSelectedPortfolio}
+            variant="compact"
           />
         ))}
       <AddNewAssetPortfolioCard
