+++ src/pages/portfolio.tsx
@@ -1,6 +1,9 @@
-import { Tabs } from "@mantine/core";
 import dynamic from "next/dynamic";
 import { FC, useEffect, useState } from "react";
+import { useRouter } from "next/router";
+import { useSearchParams } from "next/navigation";
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
+import PortfolioSections from "@/components/portfolio-widgets/portfolio-section/portfolio-section";
 import AddNewPortfolioModal from "../components/add-new-portfolio-modal";
 import LoadingForm from "../components/exchanges-form/loading";
 import PlusIconSVG from "../components/icons/plusIconSVG";
@@ -9,29 +12,48 @@ import useModal from "../hooks/useModal";
 import DefaultLayout from "../layouts/default";
 import { AssetType } from "../lib/portfolio-utils";
 import api from "../services/create-service";
+import { createUrl } from "../utils/helper";
+import mockFetchUserPortfolioData from "../tests/mocks/mock-fetchUserPortfolio-1";
+import isEmptyDataItem from "../utils/type-gaurds";
+
+const ADD_PORTFOLIO_TAB_VALUE = "add-portfolio";
+// const generateTabsValueForPortfolioItem = portfolio => `Portfolio::${portfolio._id}`;
+const generateTabsValueForPortfolioItem = portfolio => `${portfolio._id}`;
 
 const LazyLoadPortfolio = dynamic(() => import("../layouts/portfolio"), {
   loading: LoadingForm,
   ssr: false
 });
 
-const ADD_PORTFOLIO_TAB_VALUE = "add-portfolio";
-const generateTabsValueForPortfolioItem = portfolio => `Portfolio::${portfolio._id}`;
-
 const Portfolio: FC = () => {
+  const router = useRouter();
+  const searchParams = useSearchParams();
+
+  const setSelectedPortfolio = selectedCrypto => {
+    const newParams = new URLSearchParams(searchParams.toString());
+
+    if (selectedCrypto) {
+      newParams.set("q", selectedCrypto);
+    } else {
+      newParams.delete("q");
+    }
+
+    router.push(createUrl("portfolio", newParams));
+  };
+
+  const selectedPortfolio = searchParams.get("q") || undefined;
+
   const { isShowing: showCreatePortfolioModal, toggle: toggleShowCreatePortfolioModal } =
     useModal(false);
   const { isShowing: showMaxPortfolioWarningModal, toggle: toggleShowMaxPortfolioWarningModal } =
     useModal(false);
 
   const [portfolios, setPortfolios] = useState<Array<Portfolio>>([]);
-
   const portfolioCount = portfolios.length;
   const [isPortfolioFetched, setIsPortfolioFetched] = useState(false);
 
-  const [activeTab, setActiveTab] = useState(null);
   const handleAssetTypeSelection = (assetType: AssetType) => {
-    toggleShowCreatePortfolioModal();
+    // toggleShowCreatePortfolioModal();
     portfolios.push({
       _id: new Date().getTime(),
       assetType,
@@ -52,20 +74,23 @@ const Portfolio: FC = () => {
       }
       toggleShowCreatePortfolioModal();
     } else {
-      setActiveTab(tabValue);
+      // setActiveTab(tabValue);
+      setSelectedPortfolio(tabValue);
     }
   };
 
   useEffect(() => {
     (async () => {
       try {
-        const fetchUserPortfoliosResponse = await api.fetchUserPortfolio();
-        const { portfolios: userPortfolios } = fetchUserPortfoliosResponse.data;
+        // const fetchUserPortfoliosResponse = await api.fetchUserPortfolio();
+        // const { portfolios: userPortfolios } = fetchUserPortfoliosResponse.data;
 
-        const selectedPortfolioValue =
-          userPortfolios.length > 0 ? generateTabsValueForPortfolioItem(userPortfolios[0]) : null;
+        const userPortfolios = mockFetchUserPortfolioData;
 
-        setActiveTab(selectedPortfolioValue);
+        // const selectedPortfolioValue =
+        //   userPortfolios.length > 0 ? generateTabsValueForPortfolioItem(userPortfolios[0]) : null;
+
+        // setActiveTab(selectedPortfolioValue);
         setPortfolios(userPortfolios);
       } catch (error) {
         console.error(error);
@@ -82,59 +107,44 @@ const Portfolio: FC = () => {
 
   if (!isPortfolioFetched) {
     Content = <p>Loading</p>;
-  } else if (portfolioCount < 1) {
+  } else if (!isEmptyDataItem(selectedPortfolio)) {
     Content = (
-      <div className="py-4 text-center ">
-        <h1 className="font-heading mb-6 text-4xl font-bold leading-none md:text-6xl lg:text-8xl">
-          Welcome to your portfolio page!
-        </h1>
-        <p className="mb-11 text-lg font-medium text-gray-600">
-          It looks like you don&apos;t have any portfolios yet. Don&apos;t worry, you can easily
-          create a new portfolio by clicking on the{" "}
-          <span className="font-semibold text-yellow-600">&apos;Create New Portfolio&apos;</span>{" "}
-          button below. Once you have created a portfolio, you can track your investments, view your
-          estimated portfolio value, and monitor your progress. Start building your portfolio today!
-        </p>
-        <div className="w-max rounded-lg bg-charcoal-900 p-2">
-          <button
-            type="button"
-            className="rounded-lg bg-charcoal-400 p-4 font-mono font-light text-gray-500 hover:bg-yellow-700 hover:text-white focus:text-white focus:outline-none focus:ring-2 focus:ring-yellow-600 focus:ring-opacity-50"
-            onClick={() => toggleShowCreatePortfolioModal()}
+      <Tabs
+        defaultValue={selectedPortfolio}
+        onValueChange={handleTabChange}
+        className="flex h-full flex-col"
+      >
+        <TabsList className="w-fit">
+          {portfolios.map((portfolio, key) => (
+            <TabsTrigger key={portfolio._id} value={generateTabsValueForPortfolioItem(portfolio)}>
+              {`Portfolio ${key}`}
+            </TabsTrigger>
+          ))}
+          <TabsTrigger value={ADD_PORTFOLIO_TAB_VALUE}>
+            <PlusIconSVG width={15} height={15} />
+          </TabsTrigger>
+        </TabsList>
+
+        {portfolios.map(portfolio => (
+          <TabsContent
+            key={portfolio._id}
+            value={generateTabsValueForPortfolioItem(portfolio)}
+            className="flex-1"
           >
-            Create New Portfolio
-          </button>
-        </div>
-      </div>
+            {selectedPortfolio === generateTabsValueForPortfolioItem(portfolio) ? (
+              <LazyLoadPortfolio portfolio={portfolio} />
+            ) : null}
+          </TabsContent>
+        ))}
+      </Tabs>
     );
   } else {
     Content = (
-      <Tabs
-        defaultValue={activeTab}
-        onTabChange={handleTabChange}
-        // style={{ display: "flex" }}
-        // classNames={{ root: "flex flex-col flex-1", body: "flex-grow" }}
-      >
-        <Tabs.List>
-          {portfolios.map((portfolio, key) => (
-            <Tabs.Tab key={portfolio._id} value={generateTabsValueForPortfolioItem(portfolio)}>
-              {`Portfolio ${key}`}
-            </Tabs.Tab>
-          ))}
-          <Tabs.Tab value={ADD_PORTFOLIO_TAB_VALUE} icon={<PlusIconSVG width={15} height={15} />} />
-        </Tabs.List>
-
-        {portfolios.map(portfolio => (
-          <Tabs.Panel
-            key={portfolio._id}
-            value={generateTabsValueForPortfolioItem(portfolio)}
-            // classNames={{ root: "flex" }}
-          >
-            {activeTab === generateTabsValueForPortfolioItem(portfolio) ? (
-              <LazyLoadPortfolio portfolio={portfolio} />
-            ) : null}
-          </Tabs.Panel>
-        ))}
-      </Tabs>
+      <PortfolioSections
+        portfolios={portfolios}
+        setSelectedPortfolio={setSelectedPortfolio}
+        handleAssetTypeSelection={handleAssetTypeSelection}
+      />
     );
   }
 
@@ -146,15 +156,19 @@ const Portfolio: FC = () => {
         sidebar="portfolio"
         description="You can see your portfolios estimated value & progress below"
       >
-        <AddNewPortfolioModal
-          isShowing={showCreatePortfolioModal}
-          cancel={toggleShowCreatePortfolioModal}
-          onSelection={handleAssetTypeSelection}
-        />
-        <MaxPortfolioReachedModal
-          isShowing={showMaxPortfolioWarningModal}
-          cancel={toggleShowMaxPortfolioWarningModal}
-        />
+        {showCreatePortfolioModal ? (
+          <AddNewPortfolioModal
+            isShowing={showCreatePortfolioModal}
+            cancel={toggleShowCreatePortfolioModal}
+            onSelection={handleAssetTypeSelection}
+          />
+        ) : null}
+        {showMaxPortfolioWarningModal ? (
+          <MaxPortfolioReachedModal
+            isShowing={showMaxPortfolioWarningModal}
+            cancel={toggleShowMaxPortfolioWarningModal}
+          />
+        ) : null}
         <div className="portfolio-primary-panel flex h-full flex-1 flex-col overflow-y-auto rounded-lg sm:mb-1">
           {Content}
         </div>
